[api] refactoring
