+++ spl/api/__init__.py
@@ -2,6 +2,7 @@
 
 from spl.api import discretization
 from spl.api import essential_bc
-from spl.api import codegen
+from spl.api import ast
+from spl.api import printing
 from spl.api import settings
 from spl.api import grid
