+++ spl/api/printing/__init__.py
@@ -1,0 +1,1 @@
+from spl.api.printing import pycode
